add_editable_cell_renderer {gWidgets2RGtk2}	R Documentation

make a view column for the given type of variable

Description

make a view column for the given type of variable

Usage


add_editable_cell_renderer(x, self, model_idx, view_col)

Arguments

x
variable

self
reference to ModelView object

model_idx
1-based index in model of data to represent

view_col
optional view column to reuse

Value

a gtkTreeViewColumn with data "n", and "id" to block editable id

[Package gWidgets2RGtk2 version 1.0-7.1 Index]