gWidgets2RGtk2-package {gWidgets2RGtk2}	R Documentation

gWidgets2Rtk2

Description

The gWidgets2 package provides a programming interface for making graphical user interfaces within R. The package is a rewrite of the gWidgets package, introducing a few external changes but a significant number of internal ones. The package relies on one of several underlying toolkit packages providing access to the graphical libraries. This package provides the implementation to interface with the underlying RGtk2 package.

GComponent is a parent class for both GContainer and GWidget and inherits its primary interface from gWidgets2::BasicToolkitInterface.

GContainer is the base class for container objects. The main method is add_child, but there is also book-keepingn code to keep track of the child components of the container

GWidget is the Base class for widget objects

GWidgetWithItems is Base class for selection widgets based on a set of items. The key here is the handlers apply to each item, but the handler is assigned to the class member.

For RGtk2, the GButton class has the extra reference method set_border. The border argument has been deprecated.

This GTable class for RGtk2 implements a few additional reference methods: hide_names to hide the header names; remove_popup_menu to remove the popup menu; add_popup to add a popup menu

Passing in a value freeze_attributes = TRUE will make it so there are no menu items to resize frame, change variable types, relabel factors, .... Values of "row" or "column" will remove popup menus just for the row or columns.

The label class for RGtk2 has the extra method set_angle to set the angle for the text. A value of 90 will rotate ccw to read bottom to top

For RGtk2, the Gprogressbar class has the extra reference method set_border. The border argument has been deprecated.

The main reference methods GTimer are start_timer and stop_timer

Arguments

...
passed to constructor

Methods

add_drag_motion(handler, action = NULL, ...)

Called when motion over widget occurs

add_to_parent(parent, child, expand = NULL, fill = NULL, anchor = NULL, ...)

Add a child to parent if it is ia container and non null. Dispatches to add_child method of parent

get_enabled()

is widget sensistive to user input

set_enabled(value, ...)

specify with logical if widget is sensistive to user input

set_rgtk2_font(obj, value)

Set font on a gtkWidget instance

set_size(value, ...)

Set widget size (size request), value=c(width=-1, height=-1)

child_bookkeeping(child)

Update parent property of child and children property of parent container

set_child_align(child, alt_child, anchor)

Set child alignment, if a GtkMisc or GtkAlignment object

set_child_fill(child, fill, horizontal = TRUE)

Fill can be NULL, TRUE, FALSE, ”, 'both', 'x', 'y'...

connect_to_toolkit_signal(signal, decorator, emitter = .self$handler_widget())

Connect signal of toolkit to notify observer

get_value(drop = NULL, ...)

Get main value of widget. From 'svalue' when index = FALSE or NULL

remove_border()

Remove border by setting relief to none

set_value(value, ..., drop = NULL)

for 'svalue<-' when index = FALSE or NULL

add_popup(menu_fun = NULL)

Add a popup menu to the columns. Function should generate list of actions, ...

clear_columns()

Clear out old treeview columns in preparation of new

default_popup_menu(col_index)

Provide default popup menu (passed to gmenu(..., popup=TRUE))

get_index(drop = NULL, ...)

svalue; index=TRUE

get_length(...)

Get length of object. Needed for sapply.

get_model()

Helper: get rGtkDataFrame model, which is filtered

get_selected()

Get selected indices or numeric(0)

get_valid_columns()

Helper: get column indices less those for icons, tooltips, visible

get_value(drop = NULL, ...)

Get main value of widget. From 'svalue' when index = FALSE or NULL

get_view_columns()

Helper: get non-icon columns to iterate over

hide_names(value)

adjust visibility of header

make_columns()

Make new columns, watching out for icons, tooltips, visible

make_icon_column()

Make column for icons

not_these()

Helper: Remove these indices due to icon_col, ...

remove_popup_menu()

remove popup menu from column headers

set_focus(value)

If we can focus, do so, then raise

set_selected(ind)

Set selected rows by index

set_selection_mode(mode = c("none", "single", "browse", "multiple", "extended"))

Helper: Set the selection mode

set_size(value, ...)

Set widget size (size request), value=c(width=-1, height=-1)

set_value(value, ..., drop = NULL)

for 'svalue<-' when index = FALSE or NULL

get_value(drop = NULL, ...)

Get main value of widget. From 'svalue' when index = FALSE or NULL

set_angle(angle)

Rotate text by angle degrees ccw

set_value(value, ..., drop = NULL)

for 'svalue<-' when index = FALSE or NULL

get_value(drop = NULL, ...)

Get main value of widget. From 'svalue' when index = FALSE or NULL

set_value(value, ..., drop = NULL)

for 'svalue<-' when index = FALSE or NULL

set_interval(ms)

Set the interval. Need to stop and start active timer to implement.

start_timer()

Start the timer

stop_timer()

stop the timer

Author(s)

John Verzani jverzani@gmail.com

Maintainer: John Verzani jverzani@gmail.com

[Package gWidgets2RGtk2 version 1.0-7.1 Index]