getWidget.RGtkObject {gWidgets2RGtk2}	R Documentation

method for stopping getWidget

Description

method for stopping getWidget

S3 method for stopping getBlock

The gWidgets separation of API from toolkit requires the exporting of "dot" functions from each toolkit package. These are not documented, as they mirror the definition in the API package, but must be to get pass R's package check. Hence this stub.

toolkit implementation for gmessage

toolkit implementation for gconfirm

toolkit implmentation of ginput

toolkit implementation

toolkit implementation of galert

Toolkit constructor

toolkit constructor

Toolkit constructor

Toolkit XXX constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit gedit constructor

toolkit constructor for ggroup

gframe constructor

toolkit constructor

Toolkit implementation

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit label constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit constructor

Toolkit XXX constructor

Toolkit constructor

Toolkit constructor

toolkit implementation

S3 method for gtimer

Toolkit constructor

Toolkit constructor

Toolkit constructor

toolkit constructor for gwindow

add stock icons

Returns list of stock ids

return stock id

helper function

return stock id from object

Usage


## S3 method for class 'RGtkObject'
getWidget(obj)

## S3 method for class 'RGtkObject'
getBlock(obj)

.docstub()

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gmessage(toolkit, msg, title = "message",
  icon = c("info", "warning", "error", "question"), parent = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gconfirm(toolkit, msg, title = "Confirm",
  icon = c("info", "warning", "error", "question"), parent = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.ginput(toolkit, msg, text = "",
  title = "Input", icon = c("info", "warning", "error", "question"),
  parent = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gbasicdialog(toolkit, title = "Dialog",
  parent = NULL, do.buttons = TRUE, handler = NULL, action = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.galert(toolkit, msg, title = "message",
  delay = 3, parent = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gaction(toolkit, label, tooltip = NULL,
  icon = NULL, key.accel = NULL, handler = NULL, action = NULL,
  parent = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gbutton(toolkit, text, handler, action,
  container, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gcalendar(toolkit, text = "",
  format = "%Y-%m-%d", handler = NULL, action = NULL,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gcheckbox(toolkit, text, checked = FALSE,
  use.togglebutton = FALSE, handler = NULL, action = NULL,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gcheckboxgroup(toolkit = NULL, items,
  checked = FALSE, horizontal = FALSE, use.table = FALSE,
  handler = NULL, action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gcombobox(toolkit, items, selected = 1,
  editable = FALSE, coerce.with = NULL, handler = NULL, action = NULL,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gtable(toolkit, items, multiple = FALSE,
  chosen.col = 1, icon.col = NULL, tooltip.col = NULL, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gmenu(toolkit, menu.list = list(),
  popup = FALSE, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gdf(toolkit, items = NULL, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gedit(toolkit, text = "", width = 25,
  coerce.with = NULL, initial.msg = initial.msg, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.ggroup(toolkit, horizontal = TRUE,
  spacing = 5, use.scrollwindow = FALSE, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gframe(toolkit, text, markup, pos,
  horizontal = TRUE, spacing = 5, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gexpandgroup(toolkit, text, markup,
  horizontal = TRUE, handler = NULL, action = NULL, container = NULL,
  ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gfile(toolkit, text = "", type = c("open",
  "save", "selectdir"), initial.filename = NULL, initial.dir = getwd(),
  filter = list(), multi = FALSE, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gfilebrowse(toolkit, text = "",
  type = c("open", "save", "selectdir"), initial.filename = NULL,
  initial.dir = getwd(), filter = list(), quote = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gformlayout(toolkit, align = "left",
  spacing = 5, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.ggraphics(toolkit, width = dpi * 6,
  height = dpi * 6, dpi = 75, ps = 12, handler = NULL, action = NULL,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gimage(toolkit, filename = "",
  dirname = "", stock.id = NULL, size = "", handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.glabel(toolkit, text = "", markup = FALSE,
  editable = FALSE, handler = NULL, action = NULL, container = NULL,
  ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.glayout(toolkit, homogeneous = FALSE,
  spacing = 10, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gnotebook(toolkit, tab.pos = 3,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gpanedgroup(toolkit, horizontal = TRUE,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gprogressbar(toolkit, value, container, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gradio(toolkit, items, selected = 1,
  horizontal = FALSE, handler = NULL, action = NULL, container = NULL,
  ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gseparator(toolkit, horizontal = TRUE,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gslider(toolkit, from = 0, to = 100,
  by = 1, value = from, horizontal = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gspinbutton(toolkit, from = 0, to = 10,
  by = 1, value = from, digits = 0, handler = NULL, action = NULL,
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gstackwidget(toolkit, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gstatusbar(toolkit, text = "",
  container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gtext(toolkit, text = NULL, width = NULL,
  height = 300, font.attr = NULL, wrap = TRUE, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gtimer(toolkit, ms, FUN, data = NULL,
  one.shot = FALSE, start = TRUE)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gtoolbar(toolkit, toolbar.list = list(),
  style = c("both", "icons", "text", "both-horiz"), container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gtree(toolkit, offspring = NULL,
  offspring.data = NULL, chosen.col = 1, offspring.col = 2,
  icon.col = NULL, tooltip.col = NULL, multiple = FALSE, handler = NULL,
  action = NULL, container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gvarbrowser(toolkit, handler = NULL,
  action = "summary", container = NULL, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.gwindow(toolkit, title, visible = visible,
  name, width, height, parent, handler, action, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.addStockIcons(toolkit, iconNames, iconFiles,
  ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.getStockIcons(toolkit, ...)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.getStockIconByName(toolkit,name, ...)

addToGtkStockIcons(iconNames, iconFiles)

## S3 method for class 'guiWidgetsToolkitRGtk2'
.stockIconFromObject(toolkit,obj, ...)

Arguments

obj
R object to get icon from

toolkit
undocumented argument, cf. gWidgets2 documentation.

msg
undocumented argument, cf. gWidgets2 documentation.

title
undocumented argument, cf. gWidgets2 documentation.

icon
undocumented argument, cf. gWidgets2 documentation.

parent
undocumented argument, cf. gWidgets2 documentation.

...
undocumented argument, cf. gWidgets2 documentation.

text
undocumented argument, cf. gWidgets2 documentation.

do.buttons
undocumented argument, cf. gWidgets2 documentation.

handler
undocumented argument, cf. gWidgets2 documentation.

action
undocumented argument, cf. gWidgets2 documentation.

delay
undocumented argument, cf. gWidgets2 documentation.

label
undocumented argument, cf. gWidgets2 documentation.

tooltip
undocumented argument, cf. gWidgets2 documentation.

key.accel
undocumented argument, cf. gWidgets2 documentation.

container
undocumented argument, cf. gWidgets2 documentation.

format
undocumented argument, cf. gWidgets2 documentation.

checked
undocumented argument, cf. gWidgets2 documentation.

use.togglebutton
undocumented argument, cf. gWidgets2 documentation.

items
undocumented argument, cf. gWidgets2 documentation.

horizontal
undocumented argument, cf. gWidgets2 documentation.

use.table
undocumented argument, cf. gWidgets2 documentation.

selected
undocumented argument, cf. gWidgets2 documentation.

editable
undocumented argument, cf. gWidgets2 documentation.

coerce.with
undocumented argument, cf. gWidgets2 documentation.

multiple
undocumented argument, cf. gWidgets2 documentation.

chosen.col
undocumented argument, cf. gWidgets2 documentation.

icon.col
undocumented argument, cf. gWidgets2 documentation.

tooltip.col
undocumented argument, cf. gWidgets2 documentation.

menu.list
undocumented argument, cf. gWidgets2 documentation.

popup
undocumented argument, cf. gWidgets2 documentation.

width
width of device (pixels)

initial.msg
undocumented argument, cf. gWidgets2 documentation.

spacing
undocumented argument, cf. gWidgets2 documentation.

use.scrollwindow
undocumented argument, cf. gWidgets2 documentation.

markup
undocumented argument, cf. gWidgets2 documentation.

pos
undocumented argument, cf. gWidgets2 documentation.

type
undocumented argument, cf. gWidgets2 documentation.

initial.filename
undocumented argument, cf. gWidgets2 documentation.

initial.dir
undocumented argument, cf. gWidgets2 documentation.

filter
undocumented argument, cf. gWidgets2 documentation.

multi
undocumented argument, cf. gWidgets2 documentation.

quote
undocumented argument, cf. gWidgets2 documentation.

align
undocumented argument, cf. gWidgets2 documentation.

height
height of device (pixels)

dpi
dots per inch

ps
pointsize

filename
undocumented argument, cf. gWidgets2 documentation.

dirname
undocumented argument, cf. gWidgets2 documentation.

stock.id
undocumented argument, cf. gWidgets2 documentation.

size
undocumented argument, cf. gWidgets2 documentation.

homogeneous
undocumented argument, cf. gWidgets2 documentation.

tab.pos
undocumented argument, cf. gWidgets2 documentation.

value
undocumented argument, cf. gWidgets2 documentation.

from
undocumented argument, cf. gWidgets2 documentation.

to
undocumented argument, cf. gWidgets2 documentation.

by
undocumented argument, cf. gWidgets2 documentation.

digits
digits

font.attr
undocumented argument, cf. gWidgets2 documentation.

wrap
undocumented argument, cf. gWidgets2 documentation.

ms
undocumented argument, cf. gWidgets2 documentation.

FUN
undocumented argument, cf. gWidgets2 documentation.

data
undocumented argument, cf. gWidgets2 documentation.

one.shot
undocumented argument, cf. gWidgets2 documentation.

start
undocumented argument, cf. gWidgets2 documentation.

toolbar.list
undocumented argument, cf. gWidgets2 documentation.

style
undocumented argument, cf. gWidgets2 documentation.

offspring
undocumented argument, cf. gWidgets2 documentation.

offspring.data
undocumented argument, cf. gWidgets2 documentation.

offspring.col
undocumented argument, cf. gWidgets2 documentation.

visible
undocumented argument, cf. gWidgets2 documentation.

name
name of icon

iconNames
undocumented argument, cf. gWidgets2 documentation.

iconFiles
undocumented argument, cf. gWidgets2 documentation.

See Also

The documentation for this is found at gbutton.

[Package gWidgets2RGtk2 version 1.0-7.1 Index]